requirements.txt
black==19.10b0  # https://github.com/ambv/black
pre-commit==2.4.0  # https://github.com/pre-commit/pre-commit
django-debug-toolbar==2.2  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.2.9  # https://github.com/django-extensions/django-extensions
pytest-django==3.9.0  # https://github.com/pytest-dev/pytest-django
drf-yasg==1.17.1
rest-condition==1.0.3
pythonanywhere==0.9.6
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
django==3.0.7  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
# django-redis==4.11.0  # https://github.com/jazzband/django-redis
django-imagekit==4.0.2
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==4.4.0 
django-guardian==2.3.0
django-lifecycle==0.8.0
django-cors-headers==3.6.0
mysqlclient==2.0.1 # https://github.com/PyMySQL/mysqlclient-python